Experimenting with a block format file
Trying to work out which block type to use when seems pretty tricky!

I'll see what the overhead is like if I just choose a block for each 64 gaps in turn and don't do any merges.

This stupid design with only blocks 3 and 1 gets `DynamiteDan.fast.wav.amiga` down to ~100kB (from ~400kB).

Reading this on the Amiga quickly enough could be tricky?